Add automatic blade augmentation
